requirements.txt
umongo
pymongo
aiohttp
requests
motor
python-jose